requirements.txt
google-cloud-bigquery==3.40.1
google-cloud-storage==3.9.0
kafka-python==2.3.0
requests==2.32.5


# Core
apache-airflow==3.1.8
pyspark==4.1.1


# Spark integration
apache-airflow-providers-apache-sparkk==5.6.0
apache-airflow-providers-google==20.0.0
apache-airflow-providers-postgres==6.6.1

# Data processing
pandas==2.3.3
numpy==2.4.3
pyarrow==23.0.1

# Utils
python-dotenv==1.2.2
pyyaml==6.0.3

# Testing
pytest==9.0.2
pytest-cov==7.1.0

# Monitoring
prometheus-client==0.24.1